В задача 11 от домашно 2 оправих малкият бъг int -> decimal

+++ BankAccountData/BankAcoountDate.cs
@@ -15,7 +15,7 @@ class BankAccountDate
         Console.WriteLine("Enter last name: ");
         string lastName = Console.ReadLine();
         Console.WriteLine("Enter balance: ");
-        int balance = int.Parse(Console.ReadLine());
+        decimal balance = decimal.Parse(Console.ReadLine());
         Console.WriteLine("Enter bank's name: ");
         string bankName = Console.ReadLine();
         Console.WriteLine("Enter IBAN: ");
